Add CurrentUserContext to UserAuthDisplay
The goal of this commit is to cause use login/registration
to immediately update the currentUser directly,
rather than waiting for a re-render and re-fetch from
session to trigger.

This goal was working upon logout, where logout
immediately causes the Login/Logout button to swap.

But adding the Context to UserAuth is correctly updating
currentUser using the associated setter, but it is not
changing the initial state of displayLogout in the
RightNav component.
