+++ app/javascript/components/RightNav.jsx
@@ -1,43 +1,26 @@
-import React, {useContext} from "react";
+import React, {useContext, useState} from "react";
 import { Link, Redirect } from "react-router-dom"
+import AuthButton from "./AuthButton"
+
 import { CurrentUserContext } from "../contexts/CurrentUserContext"
 
 export default function RightNav(props) {
-  const {setCurrentUser} = useContext(CurrentUserContext);
-
-  const handleClick = () => {
-    const currentUserToken = sessionStorage.getItem('currentUserToken')
-    if (currentUserToken) {
-      const url = `/api/v1/session?session_token=${currentUserToken}`
-      // sessionStorage.removeItem('currentUserToken')
-      // setCurrentUser({})
-      fetch(url, {
-        method: "DELETE",
-      })
-        .then((data) => {
-          if (data.ok) {
-            sessionStorage.removeItem('currentUserToken')
-            setCurrentUser({})
-            alert("logout completed")
-            return data.json()
-          } else {
-            throw new Error("network and/or server error")
-          }
-        })
-        .catch(err => console.error("unkonwn error " + err))
-    } else {
-      alert('no currentSessionToken; no user to logout!')
-    }
-  }
+  const {currentUser} = useContext(CurrentUserContext);
+  const [displayLogout, setDisplayLogout] = useState(currentUser && currentUser.username);
+  // the initial state of currentUser is null, so the initial state
+  // of displayLogout is also null. the context also has a setter setCurrentUser.
+  //
+  // the login form (components/UserAuth/UserAuthDisplay)
+  // correctly uses this setter to update the current user
+  // upon successful login, and the RightNav component correctly tracks that change,
+  // but it isn't triggering a change in displayLogout, hence
+  // the AuthButton isn't receiving displayLogout.
 
   return (
     <nav className='right-nav'>
       <ul className = "flex-container-row">
         <li>
-          <Link to="/session/new">Log In</Link>
-        </li>
-        <li>
-          <a onClick = {handleClick}>Log Out</a>
+          <AuthButton displayLogout = {displayLogout} setDisplayLogout = {setDisplayLogout}/>
         </li>
       </ul>
     </nav>
